Move links to top of launcher and fix changelog entries not being aligned
